Add notes of stan-functions.txt (stuck in 2.19.0)
As of now, rstan is still using stan 2.19.
"But we haven’t been able to get rstan beyond the Stan 2.19 branch, no so much due to stanc3 but due to the CRTP and some non-deterministic compiler behavior."
https://discourse.mc-stan.org/t/cmdstanr-and-rstudio-stan-syntax-highlighting/13072/6

The automatic generation of stan-functions.txt from the latest stan requires the following be merged.
https://github.com/stan-dev/docs/pull/128